feat: Improved sliceConcat inPlace determiniation
Prior we would exclude any views being used. However since the first slice is the only slice appended into, it's the only one we care if it's a view.
This improved logic allows more in place operations when multiple slices are used and the first slice does not result in a view..